Introduce key parsers and refactor normal mode input with them
We now have a quick DFA-like language for reflecting sequences of keys, which
cuts a lot of the conditional logic out of key handlers.

We've refactored normal mode in terms of these, which brings us most of the way
to refactoring everything (more to come).

In support of this change we've also taken the buf and sline arguments away
from the old key handler and are replacing them with methods on ui().

Signed-off-by: Casey Dahlin <casey.dahlin@gmail.com>